Refactor code structure and
Improves accuracy of time step and index

Corrects the calculation of the alpha parameter, which represents the fraction of the hour, to start from the beginning of the time step (k+1).
Fixes an off-by-one error in the output index calculation, ensuring correct assignment of values to the output arrays.

Ensures same sign convention for cooling power in validation script.